fixes added

+++ SystemExpertANN/DataAccessLayer/Becarios.cs
@@ -21,7 +21,7 @@ namespace DataAccessLayer
         public int Edad { get; set; }
         public double Promedio { get; set; }
         public bool EsRegular { get; set; }
-        public double IngresoMensual { get; set; }
+        public float IngresoMensual { get; set; }
         public bool EsBecadoProspera { get; set; }
         public bool Discapacidad { get; set; }
         public int Id_Municipio { get; set; }
